requirements.txt
absl-py==2.1.0
astunparse==1.6.3
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
dacite==1.8.1
filelock==3.20.3
Flask==3.0.1
Flask-MySQLdb==2.0.0
flatbuffers==23.5.26
fonttools==4.40.0
fsspec==2026.1.0
gast==0.5.4
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.0
h11==0.16.0
h5py==3.10.0
htmlmin==0.1.12
idna==3.4
ImageHash==4.3.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
keras==2.15.0
kiwisolver==1.4.4
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.3
matplotlib==3.7.1
ml-dtypes==0.2.0
mpmath==1.3.0
multimethod==1.9.1
mysqlclient==2.2.1
networkx==3.1
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
patsy==0.5.3
phik==0.12.3
Pillow==9.5.0
protobuf==4.23.4
pyarrow==12.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.10.10
pylance==0.5.2
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sklearn==0.0.post5
statsmodels==0.14.0
sympy==1.14.0
tangled-up-in-unicode==0.2.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.1.0
torch==2.9.1
tqdm==4.65.0
typeguard==2.13.3
typing_extensions==4.15.0
tzdata==2023.3
urllib3==2.0.3
uvicorn==0.40.0
visions==0.7.5
Werkzeug==3.0.1
wordcloud==1.9.2
wrapt==1.14.1
ydata-profiling==4.3.1
